Renamed ValidatingTextField to DetailField. Moved validation into its own class TrialValidator. Changed ValidationError to only be able to be created by TrialValidator to prevent validation from occurring elsewhere.
